requirements.txt
google-cloud~=0.34.0
google-cloud-logging~=2.7.0
google-cloud-error-reporting~=1.5.3
